Added collider, clouds, rain, fire, falgs, island and sea.
The game character cannot leave the environment or in our case, go beyond the environment into the sea.
I have also added a rain effect with a stormy cloud on top that emits rain.
I have created an island and a sea in between the game environment and the Island.
Also, added a torch effect and flags for the levels.